fix(math): own the bounded Gröbner worker in its external-tool module
The architecture gateway requires run_bounded_process callers to be
registered concrete external-tool owners, mirroring the Singular
adapter.  Move the killable complete-kernel strategy into
polynomials/_groebner_worker.py and register it.
