+++ tools/check_architecture.py
@@ -20,6 +20,7 @@ _EXTERNAL_OPERATION_OWNERS = frozenset(
         _PROCESS_OWNER,
         PurePosixPath("src/jacobian/math/commutative_algebra_ops/_singular.py"),
         PurePosixPath("src/jacobian/math/logic/_operations.py"),
+        PurePosixPath("src/jacobian/math/polynomials/_groebner_worker.py"),
     }
 )
 _GENERATED_DIRECTORIES = frozenset(
